requirements.txt
google-generativeai
python-dotenv
chromadb
pillow
pymupdf
pymupdf4llm
langchain-text-splitters
rich
markdown